+++ examples/listen-ticks.js
@@ -1,10 +1,10 @@
-import { connect, marketWatcher, } from "@reiryoku/mida";
+import { connect, watchTicks, } from "@reiryoku/mida";
 
 /*
  * Listen real-time Bitcoin price updates
  */
 
-// More account login examples at https://www.mida.org/documentation/essentials/login.html
+// More account login examples at https://www.mida.org/documentation/essentials/connect.html
 const myAccount = await connect("cTrader", {
     clientId: "***",
     clientSecret: "***",
@@ -12,14 +12,8 @@ const myAccount = await connect("cTrader", {
     accountId: "***",
 });
 
-const watcher = await marketWatcher({ tradingAccount: myAccount, });
-
-watcher.on("tick", (e) => {
+watchTicks(myAccount, "BTCUSDT", (e) => {
     const { tick, } = e.descriptor;
 
     console.log(`Bitcoin price is now ${tick.bid}$`);
-});
-
-await watcher.watch("BTCUSD", {
-    watchTicks: true,
 });